Filter a XenaHub Object

Major function of UCSCXenatools. It is used to filter XenaHub object according to cohorts, datasets. All datasets can be found at <https://xenabrowser.net/datapages/>. Note, the change for filtering cohorts and datasets are independent.

XenaFilter(x, filterCohorts = NULL, filterDatasets = NULL,
  ignore.case = TRUE)

Arguments

x
a XenaHub object

filterCohorts
default is NULL. A character used to filter cohorts, regular expression is supported.

filterDatasets
default is NULL. A character used to filter datasets, regular expression is supported.

ignore.case
if FALSE, the pattern matching is case sensitive and if TRUE, case is ignored during matching.

Value

a XenaHub object

Examples

# operate TCGA datasets
xe = XenaGenerate(subset = XenaHostNames == "TCGA")
xe

#> class: XenaHub 
#> hosts():
#>   https://tcga.xenahubs.net
#> cohorts() (38 total):
#>   TCGA Acute Myeloid Leukemia (LAML)
#>   TCGA Adrenocortical Cancer (ACC)
#>   TCGA Bile Duct Cancer (CHOL)
#>   ...
#>   TCGA Thyroid Cancer (THCA)
#>   TCGA Uterine Carcinosarcoma (UCS)
#> datasets() (879 total):
#>   TCGA.LAML.sampleMap/HumanMethylation27
#>   TCGA.LAML.sampleMap/HumanMethylation450
#>   TCGA.LAML.sampleMap/Gistic2_CopyNumber_Gistic2_all_data_by_genes
#>   ...
#>   TCGA.UCS.sampleMap/Pathway_Paradigm_RNASeq_And_Copy_Number
#>   TCGA.UCS.sampleMap/mutation_curated_broad

# get all names of clinical data
xe2 = XenaFilter(xe, filterDatasets = "clinical")
datasets(xe2)

#>  [1] "TCGA.OV.sampleMap/OV_clinicalMatrix"            
#>  [2] "TCGA.DLBC.sampleMap/DLBC_clinicalMatrix"        
#>  [3] "TCGA.KIRC.sampleMap/KIRC_clinicalMatrix"        
#>  [4] "TCGA.SARC.sampleMap/SARC_clinicalMatrix"        
#>  [5] "TCGA.COAD.sampleMap/COAD_clinicalMatrix"        
#>  [6] "TCGA.PRAD.sampleMap/PRAD_clinicalMatrix"        
#>  [7] "TCGA.LUSC.sampleMap/LUSC_clinicalMatrix"        
#>  [8] "TCGA.ACC.sampleMap/ACC_clinicalMatrix"          
#>  [9] "TCGA.KICH.sampleMap/KICH_clinicalMatrix"        
#> [10] "TCGA.UCS.sampleMap/UCS_clinicalMatrix"          
#> [11] "TCGA.COADREAD.sampleMap/COADREAD_clinicalMatrix"
#> [12] "TCGA.LUNG.sampleMap/LUNG_clinicalMatrix"        
#> [13] "TCGA.LUAD.sampleMap/LUAD_clinicalMatrix"        
#> [14] "TCGA.FPPP.sampleMap/FPPP_clinicalMatrix"        
#> [15] "TCGA.LAML.sampleMap/LAML_clinicalMatrix"        
#> [16] "TCGA.GBM.sampleMap/GBM_clinicalMatrix"          
#> [17] "TCGA.KIRP.sampleMap/KIRP_clinicalMatrix"        
#> [18] "TCGA.PAAD.sampleMap/PAAD_clinicalMatrix"        
#> [19] "TCGA.CHOL.sampleMap/CHOL_clinicalMatrix"        
#> [20] "TCGA.CESC.sampleMap/CESC_clinicalMatrix"        
#> [21] "TCGA.SKCM.sampleMap/SKCM_clinicalMatrix"        
#> [22] "TCGA.LGG.sampleMap/LGG_clinicalMatrix"          
#> [23] "TCGA.PCPG.sampleMap/PCPG_clinicalMatrix"        
#> [24] "TCGA.TGCT.sampleMap/TGCT_clinicalMatrix"        
#> [25] "TCGA.BLCA.sampleMap/BLCA_clinicalMatrix"        
#> [26] "TCGA.THYM.sampleMap/THYM_clinicalMatrix"        
#> [27] "TCGA.BRCA.sampleMap/BRCA_clinicalMatrix"        
#> [28] "TCGA.UVM.sampleMap/UVM_clinicalMatrix"          
#> [29] "TCGA.UCEC.sampleMap/UCEC_clinicalMatrix"        
#> [30] "TCGA.LIHC.sampleMap/LIHC_clinicalMatrix"        
#> [31] "TCGA.GBMLGG.sampleMap/GBMLGG_clinicalMatrix"    
#> [32] "TCGA.THCA.sampleMap/THCA_clinicalMatrix"        
#> [33] "TCGA.HNSC.sampleMap/HNSC_clinicalMatrix"        
#> [34] "TCGA.ESCA.sampleMap/ESCA_clinicalMatrix"        
#> [35] "TCGA.STAD.sampleMap/STAD_clinicalMatrix"        
#> [36] "TCGA.READ.sampleMap/READ_clinicalMatrix"        
#> [37] "TCGA.MESO.sampleMap/MESO_clinicalMatrix"