Filter a XenaHub Object

One of main functions in UCSCXenatools. It is used to filter XenaHub object according to cohorts, datasets. All datasets can be found at https://xenabrowser.net/datapages/.

XenaFilter(
  x,
  filterCohorts = NULL,
  filterDatasets = NULL,
  ignore.case = TRUE,
  ...
)

Arguments

x
a XenaHub object

filterCohorts
default is NULL. A character used to filter cohorts, regular expression is supported.

filterDatasets
default is NULL. A character used to filter datasets, regular expression is supported.

ignore.case
if FALSE, the pattern matching is case sensitive and if TRUE, case is ignored during matching.

...
other arguments except value passed to base::grep().

Value

a XenaHub object

Author

Shixiang Wang w_shixiang@163.com

Examples

# operate TCGA datasets
xe = XenaGenerate(subset = XenaHostNames == "tcgaHub")
xe
# get all names of clinical data
xe2 = XenaFilter(xe, filterDatasets = "clinical")
datasets(xe2)